Refactoring
* Move all functions for string processing into procs.rs from main.rs
* Fix some comments
